+++ generator/downloader/__init__.py
@@ -1,2 +1,3 @@
 from .fira_code import *
 from .nerd_font_patcher import *
+from .noto_devanagari import *